% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{BalanceMinority}
\alias{BalanceMinority}
\title{Re-balances target on learning data}
\usage{
BalanceMinority(learningData, level)
}
\arguments{
\item{learningData}{the data to re-balance}

\item{level}{the level of the target}
}
\description{
Re-balances target on learning data
}
