% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_file.R
\name{data_file.load_currencies}
\alias{data_file.load_currencies}
\title{Load currencies data from CSV file}
\usage{
data_file.load_currencies()
}
\value{
a data.table containing the currencies
}
\description{
Load currencies data from CSV file
}
