% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_file.R
\name{data_file.load_euribor}
\alias{data_file.load_euribor}
\title{Load EURIBOR values from CSV file}
\usage{
data_file.load_euribor()
}
\value{
a data.table containing the EURIBOR's values
}
\description{
Load EURIBOR values from CSV file
}
\examples{
data_file.load_euribor()
}
