% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_file.R
\name{data_file.load_fchi}
\alias{data_file.load_fchi}
\title{Load FCHI (CAC40) indice from CSV file}
\usage{
data_file.load_fchi()
}
\value{
a data.table containing the indice's values
}
\description{
Load FCHI (CAC40) indice from CSV file
}
\examples{
data_file.load_fchi()
}
