% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_file.R
\name{data_file.load_vix}
\alias{data_file.load_vix}
\title{Load VIX values from CSV file}
\usage{
data_file.load_vix()
}
\value{
a data.table containing the VIX's values
}
\description{
Load VIX values from CSV file
}
