% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_file.R
\name{data_file.load_vxd}
\alias{data_file.load_vxd}
\title{Load VXD values from CSV file}
\usage{
data_file.load_vxd()
}
\value{
a data.table containing the VXD's values
}
\description{
Load VXD values from CSV file
}
\examples{
data_file.load_vxd()
}
