% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{data_tools.extract_forecasting_data}
\alias{data_tools.extract_forecasting_data}
\title{Extracts the forecasting data depending on the forecast date}
\usage{
data_tools.extract_forecasting_data(x, jour_date)
}
\arguments{
\item{x}{}

\item{jour_date}{}
}
\description{
Extracts the forecasting data depending on the forecast date
}
