% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{data_tools.extract_forecasting_data_by_level}
\alias{data_tools.extract_forecasting_data_by_level}
\title{Extracts the forecasting data depending on the forecast date and the level.}
\usage{
data_tools.extract_forecasting_data_by_level(x, jour_date, level)
}
\arguments{
\item{x}{the \code{data.table} containing the forecasting data}

\item{level}{}
}
\description{
Extracts the forecasting data depending on the forecast date and the level.
}
