% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{data_tools.filter_nonnumeric}
\alias{data_tools.filter_nonnumeric}
\title{Filters non numerical columns}
\usage{
data_tools.filter_nonnumeric(x)
}
\arguments{
\item{x}{the \code{data.table} to filter.}
}
\description{
Filters non numerical columns
}
