% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{data_tools.filter_trend}
\alias{data_tools.filter_trend}
\title{Filters \code{Trend} and Target columns}
\usage{
data_tools.filter_trend(x)
}
\arguments{
\item{x}{the \code{data.table} to filter.}
}
\value{
the \code{data.table} filtered from all \code{Trend}.
}
\description{
Filters \code{Trend} and Target columns
}
