% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{data_tools.filter_trend_by_level}
\alias{data_tools.filter_trend_by_level}
\title{Filters \code{Trend} column of specific level.}
\usage{
data_tools.filter_trend_by_level(x, level)
}
\arguments{
\item{x}{the \code{data.table} to filter.}

\item{level}{the level to filter.}
}
\value{
the \code{data.table} filtered from \code{Trend} of \code{level}.
}
\description{
Filters \code{Trend} column of specific level.
}
