% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{data_tools.flat_rebalance_raw_data}
\alias{data_tools.flat_rebalance_raw_data}
\title{Flats and rebalances raw data by selecting good known flag r and define Target variable.}
\usage{
data_tools.flat_rebalance_raw_data(
  raw_data,
  level,
  day_date,
  start_flat_raw_data,
  n
)
}
\arguments{
\item{n}{the number of samples for each target}
}
\value{

}
\description{
Flats and rebalances raw data by selecting good known flag r and define Target variable.
}
