% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{data_tools.get_daily_forecasting_dates}
\alias{data_tools.get_daily_forecasting_dates}
\title{Retrieve all possible daily forecasting dates for this asset.}
\usage{
data_tools.get_daily_forecasting_dates(asset_code, last_date)
}
\arguments{
\item{asset_code}{a character vector indicating the asset. code.}

\item{last_date}{the last date of existing forecasts.}
}
\value{
a vector containing the daily forecasting dates.
}
\description{
Retrieve all possible daily forecasting dates for this asset.
}
