% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{data_tools.get_forecasting_dates}
\alias{data_tools.get_forecasting_dates}
\title{Retrieve all possible forecasting dates for this asset.}
\usage{
data_tools.get_forecasting_dates(asset_code, start_date, end_date)
}
\arguments{
\item{asset_code}{a character vector indicating the asset. code.}

\item{start_date}{the date when forecasting starts.}

\item{end_date}{the date when forecasting ends.}
}
\value{
a vector containing the forecasting dates.
}
\description{
Retrieve all possible forecasting dates for this asset.
}
