% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{data_tools.get_target}
\alias{data_tools.get_target}
\title{Builds two-class \code{Factor} target}
\usage{
data_tools.get_target(x, order)
}
\arguments{
\item{x}{the \code{data.table} containing the target classes.}

\item{order}{the meaningful order type (e.g. Constants.BUY, Constants.SELL, Constants.NEUTRAL)}
}
\value{
the two-class \code{Factor} target
}
\description{
Builds two-class \code{Factor} target
}
