% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{data_tools.sample}
\alias{data_tools.sample}
\title{Samples data to filter forecast.}
\usage{
data_tools.sample(flat_data, forecast_data, order, level, day_date)
}
\arguments{
\item{flat_data}{the data to sample}

\item{forecast_data}{the data from to sample}

\item{day_date}{}
}
\value{

}
\description{
Samples data to filter forecast.
}
