% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{data_tools.target_to_factor}
\alias{data_tools.target_to_factor}
\title{Converts target into factor}
\usage{
data_tools.target_to_factor(learning_data, order)
}
\arguments{
\item{learning_data}{the date to convert}

\item{order}{the order of conversion}
}
\value{
the converted data
}
\description{
Converts target into factor
}
