% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{data_tools.transform_target}
\alias{data_tools.transform_target}
\title{Transform Target to a 2-class factor depending on level and order.}
\usage{
data_tools.transform_target(x, order, level)
}
\arguments{
\item{level}{}
}
\value{

}
\description{
Transform Target to a 2-class factor depending on level and order.
}
