% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{data_tools.transform_to_sample}
\alias{data_tools.transform_to_sample}
\title{Transforms flat data into sample ones.}
\usage{
data_tools.transform_to_sample(x, order, level, with_geo_pt_t2)
}
\arguments{
\item{x}{the \code{data.table} to transform.}

\item{order}{the meaningful order type (e.g. Constants.ACHAT, Constants.VENTE, Constants.NEUTRE).}

\item{level}{the level to consider.}

\item{with_geo_pt_t2}{TRUE if Geometric Points T2 are used.}
}
\value{
a \code{data.table} containing the sample data.
}
\description{
Transforms flat data into sample ones.
}
