% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_asset.R
\name{db_asset.delete}
\alias{db_asset.delete}
\title{Delete specified asset from DB if exists.}
\usage{
db_asset.delete(name, to_delete_asset)
}
\arguments{
\item{name}{the name of the asset (e.g. CAC_40, BNP)}

\item{to_delete_asset}{the \code{data.table} containing the asset to delete.}
}
\value{
Nothing.
}
\description{
Delete specified asset from DB if exists.
}
