% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_asset.R
\name{db_asset.load}
\alias{db_asset.load}
\title{Load asset data from DB.}
\usage{
db_asset.load(name)
}
\arguments{
\item{name}{the name of the asset (e.g. CAC_40, BNP)}
}
\value{
a \code{data.table} containing the asset data.
}
\description{
Load asset data from DB.
}
