% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_asset.R
\name{db_asset.save}
\alias{db_asset.save}
\title{Save new asset only into DB.}
\usage{
db_asset.save(name, asset)
}
\arguments{
\item{name}{the name of the asset (e.g. CAC_40, BNP)}

\item{asset}{the \code{data.table} containing the asset data.}
}
\value{
Nothing
}
\description{
Save new asset only into DB.
}
