% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_asset.R
\name{db_asset.update}
\alias{db_asset.update}
\title{Update specified asset into DB if exists.}
\usage{
db_asset.update(name, to_update_asset)
}
\arguments{
\item{name}{the name of the asset (e.g. CAC_40, BNP)}

\item{to_update_asset}{the \code{data.table} containing the asset to update.}
}
\value{
Nothing.
}
\description{
Update specified asset into DB if exists.
}
