% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_currency.R
\name{db_currency.delete}
\alias{db_currency.delete}
\title{Delete one specified Currency from DB if exists.}
\usage{
db_currency.delete(name, to_delete_currency)
}
\arguments{
\item{name}{the name of the Currency (e.g. EUR_CHF, EUR_USD)}

\item{to_delete_currency}{the \code{data.table} containing the Currency to delete.}
}
\value{
Nothing.
}
\description{
Delete one specified Currency from DB if exists.
}
