% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_currency.R
\name{db_currency.load}
\alias{db_currency.load}
\title{Load all Currencies from DB.}
\usage{
db_currency.load()
}
\value{
a \code{data.table} containing all the Currencies.
}
\description{
Load all Currencies from DB.
}
