% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_currency.R
\name{db_currency.load_one}
\alias{db_currency.load_one}
\title{Load one Currency from DB.}
\usage{
db_currency.load_one(name)
}
\arguments{
\item{name}{the name of the Currency (e.g. EUR_CHF, EUR_USD)}
}
\value{
a \code{data.table} containing the Currency.
}
\description{
Load one Currency from DB.
}
