% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_currency.R
\name{db_currency.save}
\alias{db_currency.save}
\title{Save Currency into DB.}
\usage{
db_currency.save(name, currency)
}
\arguments{
\item{name}{the name of the Currency (e.g. EUR_CHF, EUR_USD)}

\item{currency}{\code{data.table} containing the Currency to save.}
}
\value{
Nothing.
}
\description{
Save Currency into DB.
}
