% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_ibor.R
\name{db_ibor.delete}
\alias{db_ibor.delete}
\title{Delete one specified IBOR from DB if exists.}
\usage{
db_ibor.delete(name, period, to_delete_ibor)
}
\arguments{
\item{name}{the name of the IBOR (e.g. EURIBOR, LIBOR)}

\item{period}{the period of the IBOR (e.g. 3M, 12M)}

\item{to_delete_ibor}{the \code{data.table} containing the IBOR to delete.}
}
\value{
Nothing.
}
\description{
Delete one specified IBOR from DB if exists.
}
