% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_ibor.R
\name{db_ibor.load}
\alias{db_ibor.load}
\title{Load all IBOR Index from DB.}
\usage{
db_ibor.load(name)
}
\arguments{
\item{name}{the name of the index (e.g. EURIBOR, LIBOR)}
}
\value{
a \code{data.table} containing the IBOR index.
}
\description{
Load all IBOR Index from DB.
}
