% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_ibor.R
\name{db_ibor.save}
\alias{db_ibor.save}
\title{Save IBOR Index into DB.}
\usage{
db_ibor.save(name, period, ibor)
}
\arguments{
\item{name}{the name of the IBOR (e.g. EURIBOR, LIBOR)}

\item{period}{the period of the IBOR (e.g. 3M, 12M)}

\item{ibor}{\code{data.table} containing the IBOR index to save.}
}
\value{
Nothing.
}
\description{
Save IBOR Index into DB.
}
