% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_ibor.R
\name{db_ibor.update}
\alias{db_ibor.update}
\title{Update one specified IBOR into DB if exists.}
\usage{
db_ibor.update(name, period, to_update_ibor)
}
\arguments{
\item{name}{the name of the IBOR (e.g. EURIBOR, LIBOR)}

\item{period}{the period of the IBOR (e.g. 3M, 12M)}

\item{to_update_ibor}{the \code{data.table} containing the IBOR to update.}
}
\value{
Nothing.
}
\description{
Update one specified IBOR into DB if exists.
}
