% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{db_utils.delete}
\alias{db_utils.delete}
\title{Internal optimized delete function by factorizing code.}
\usage{
db_utils.delete(to_delete_data, collection_name, document_name)
}
\arguments{
\item{to_delete_data}{the \code{data.table} containing the data to delete.}

\item{collection_name}{the name of the mongodb collection.}

\item{document_name}{the name of the documents in this collection.}
}
\value{
Nothing
}
\description{
Internal optimized delete function by factorizing code.
}
