% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{db_utils.load}
\alias{db_utils.load}
\title{Internal optimized load function by factorizing code.}
\usage{
db_utils.load(collection_name, document_name)
}
\arguments{
\item{collection_name}{the name of the mongodb collection.}

\item{document_name}{the name of the documents in this collection.}
}
\value{
a \code{data.table} containing the loaded data.
}
\description{
Internal optimized load function by factorizing code.
}
