% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{db_utils.name_column}
\alias{db_utils.name_column}
\title{Internal optimization storage by renaming column.}
\usage{
db_utils.name_column(dt, to_keep, to_name, name)
}
\arguments{
\item{dt}{the \code{data.table} to process.}

\item{to_keep}{a vector of columns to NOT process.}

\item{to_name}{a vector of columns to process.}

\item{name}{the prefix to add.}
}
\value{
the renamed columns \code{data.table}.
}
\description{
Internal optimization storage by renaming column.
}
