% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{db_utils.save}
\alias{db_utils.save}
\title{Internal optimized save function by factorizing code.}
\usage{
db_utils.save(new_data, collection_name, document_name)
}
\arguments{
\item{new_data}{the \code{data.table} containing the data to save.}

\item{collection_name}{the name of the mongodb collection.}

\item{document_name}{the name of the documents in this collection.}
}
\value{
Nothing
}
\description{
Internal optimized save function by factorizing code.
}
