% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{db_utils.update}
\alias{db_utils.update}
\title{Internal optimized update function by factorizing code.}
\usage{
db_utils.update(to_update_data, collection_name, document_name)
}
\arguments{
\item{to_update_data}{the \code{data.table} containing the data to update.}

\item{collection_name}{the name of the mongodb collection.}

\item{document_name}{the name of the documents in this collection.}
}
\value{
Nothing
}
\description{
Internal optimized update function by factorizing code.
}
