% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_volatility.R
\name{db_volatility.load}
\alias{db_volatility.load}
\title{Load Volatility Index Index from DB.}
\usage{
db_volatility.load(name)
}
\arguments{
\item{name}{the name of the index (e.g. VIX, VXD)}
}
\value{
a \code{data.table} containing the Volatility Index index.
}
\description{
Load Volatility Index Index from DB.
}
