% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_volatility.R
\name{db_volatility.save}
\alias{db_volatility.save}
\title{Save Volatility Index Index into DB.}
\usage{
db_volatility.save(name, volatility)
}
\arguments{
\item{name}{the name of the Volatility Index (e.g. VIX, VXD)}

\item{volatility}{\code{data.table} containing the Volatility Index index to save.}
}
\value{
Nothing.
}
\description{
Save Volatility Index Index into DB.
}
