% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_volatility.R
\name{db_volatility.update}
\alias{db_volatility.update}
\title{Update one specified Volatility Index into DB if exists.}
\usage{
db_volatility.update(name, to_update_volatility)
}
\arguments{
\item{name}{the name of the Volatility Index (e.g. VIX, VXD)}

\item{to_update_volatility}{the \code{data.table} containing the Volatility Index to update.}
}
\value{
Nothing.
}
\description{
Update one specified Volatility Index into DB if exists.
}
