% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_point.R
\name{delta_point.build}
\alias{delta_point.build}
\title{Builds Delta Points}
\usage{
delta_point.build(asset_data)
}
\arguments{
\item{asset_data}{}
}
\value{
a \code{data.table} containing the delta points.
}
\description{
Builds Delta Points
}
