% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_euribor.R
\name{euribor.get_pca}
\alias{euribor.get_pca}
\title{Computes PCA on EURIBOR.}
\usage{
euribor.get_pca(day_date)
}
\arguments{
\item{day_date}{a character vector indicating the date of the day until which PCA will be computed.}
}
\value{
a data.table containing the formatted result of the EURIBOR PCA.
}
\description{
Computes PCA on EURIBOR.
}
