% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_csv.R
\name{export_csv.X_y}
\alias{export_csv.X_y}
\title{Export X & y as CSV files}
\usage{
export_csv.X_y(
  asset_code = "CAC_40",
  start_date = "2022-08-15",
  end_date = "2022-11-13",
  full_backup_path
)
}
\description{
Export X & y as CSV files
}
