% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_r.R
\name{flag_r.build}
\alias{flag_r.build}
\title{Builds Flag R at once (bulk process).}
\usage{
flag_r.build(asset_data)
}
\arguments{
\item{asset_data}{a data.table containing the asset data.}
}
\value{
a \code{list} containing
\itemize{
\item{\code{value}   a \code{data.table} containing the Flag R values.}
\item{\code{ttt}     a TTT (Technical Temporary Table) used for updating.}
}
}
\description{
Builds Flag R at once (bulk process).
}
