% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_r.R
\name{flag_r.extract}
\alias{flag_r.extract}
\title{Extracts Flag R from its temporary technical table.}
\usage{
flag_r.extract(ttt_data)
}
\arguments{
\item{ttt_data}{the temporary technical table.}
}
\value{
the \code{flag_r}
}
\description{
Extracts Flag R from its temporary technical table.
}
