% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_r.R
\name{flag_r.update_daily}
\alias{flag_r.update_daily}
\title{Builds and updates R Flag daily}
\usage{
flag_r.update_daily(previous_ttt_data, day_asset_data)
}
\arguments{
\item{previous_ttt_data}{the TTT data from previous day.}

\item{day_asset_data}{asset data on one day.}
}
\value{
the updated Flag R with one more day.
}
\description{
Builds and updates R Flag daily
}
