% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_r.R
\name{flag_r_bind_table_flag}
\alias{flag_r_bind_table_flag}
\title{Binds 2 lists of \code{data.table}.}
\usage{
flag_r_bind_table_flag(table_flag1, table_flag2)
}
\arguments{
\item{table_flag1}{the list #1.}

\item{table_flag2}{the list #2.}
}
\value{
one bound list.
}
\description{
Binds 2 lists of \code{data.table}.
}
