% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_r.R
\name{flag_r_compute_flag_r_init}
\alias{flag_r_compute_flag_r_init}
\title{Computes stoploss, target, signal and neutral flags.}
\usage{
flag_r_compute_flag_r_init(level, ttt_data)
}
\arguments{
\item{level}{the variation level.}

\item{DT}{the temporary technical table (TTT).}
}
\value{
the temporary technical table (TTT) with the computed flags.
}
\description{
Computes stoploss, target, signal and neutral flags.
}
