% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_r.R
\name{flag_r_fix_flag_neutral}
\alias{flag_r_fix_flag_neutral}
\title{Computes and fixes neutral flags on daily process.}
\usage{
flag_r_fix_flag_neutral(level, new_ttt_data, previous_ttt_data)
}
\arguments{
\item{previous_ttt_data}{}
}
\description{
Computes and fixes neutral flags on daily process.
}
