% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_r.R
\name{flag_r_get_flag_neutral}
\alias{flag_r_get_flag_neutral}
\title{Computes neutral flags on bulk process.}
\usage{
flag_r_get_flag_neutral(level, DT)
}
\arguments{
\item{DT}{}
}
\description{
Computes neutral flags on bulk process.
}
