% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_r.R
\name{flag_r_initialize_objectif_stoploss}
\alias{flag_r_initialize_objectif_stoploss}
\title{Initializes the temporary technical table (TTT).}
\usage{
flag_r_initialize_objectif_stoploss(level, asset_data, seuil = 1/5)
}
\arguments{
\item{level}{the level to initialize.}

\item{seuil}{the threshold used to compute target and stoploss.}

\item{DT}{the asset data used as skeleton for the TTT.}
}
\value{
the temporary technical table (TTT).
}
\description{
Initializes the temporary technical table (TTT).
}
