% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_technic.R
\name{indicator_technic.get}
\alias{indicator_technic.get}
\title{Technical Indicator}
\usage{
indicator_technic.get(indice_code, day_date)
}
\arguments{
\item{indice_code}{a character vector indicating the indice code on which computation will be done.}

\item{day_date}{a character vector indicating the date of the day until which computation will be done.}
}
\value{
a data.table containing the technical indicators.
}
\description{
This function computes all technicals indicators.
}
