% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{ReBalanceData}
\alias{ReBalanceData}
\title{Re-balances target on learning data with \code{unbalanced::ubBalance}}
\usage{
ReBalanceData(learningData, level, type = "ubSMOTE")
}
\arguments{
\item{learningData}{the data to re-balance}

\item{level}{the level of the target}

\item{type}{the balancing technique to use (ubOver, ubUnder, ubSMOTE, ubOSS, ubCNN, ubENN, ubNCL, ubTomek).}
}
\value{

}
\description{
Re-balances target on learning data with \code{unbalanced::ubBalance}
}
