% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_tools.R
\name{technical_tools.dynamic_merge}
\alias{technical_tools.dynamic_merge}
\title{Dynamic merge of data.tables list with INNER join.}
\usage{
technical_tools.dynamic_merge(
  list_dt,
  exclude = "something_that_we_will_never_name_like_that"
)
}
\arguments{
\item{list_dt}{a list of data.tables to merge.
The list should contains at least 2 data.tables.
And they should have all the same data.table keys.}

\item{exclude}{the name of a \code{data.table} in list_dt to exclude from the merge.}
}
\value{
a data.table containing all the data.tables merged fields.
}
\description{
Dynamic merge of data.tables list with INNER join.
}
