% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_tools.R
\name{technical_tools.dynamic_merge_full}
\alias{technical_tools.dynamic_merge_full}
\title{Dynamic merge of data.tables list with FULL OUTER join.}
\usage{
technical_tools.dynamic_merge_full(list_dt)
}
\arguments{
\item{list_dt}{a list of data.tables to merge.
The list should contains at least 2 data.tables.
And they should have all the same data.table keys.}
}
\value{
a data.table containing all the data.tables merged fields.
}
\description{
Dynamic merge of data.tables list with FULL OUTER join.
}
