% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_tools.R
\name{technical_tools.subtract_year}
\alias{technical_tools.subtract_year}
\title{Subtracts period from a date.}
\usage{
technical_tools.subtract_year(a_date, year)
}
\arguments{
\item{a_date}{the date.}

\item{year}{the number of years to substract (e.g. 0.5, 1, 4)}
}
\value{
the result of the subtract.
}
\description{
Subtracts period from a date.
}
