% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{data_tools.extract_learning_data}
\alias{data_tools.extract_learning_data}
\title{Extracts the learning data depending on the forecast date}
\usage{
data_tools.extract_learning_data(x, day_date)
}
\arguments{
\item{x}{the \code{data.table} containing the learning data}

\item{day_date}{the date until when extraction is done}
}
\description{
Extracts the learning data depending on the forecast date
}
