% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{data_tools.flat_raw_data}
\alias{data_tools.flat_raw_data}
\title{Flats raw data by selecting good known flag r and define Target variable.}
\usage{
data_tools.flat_raw_data(raw_data, level, day_date, start_flat_raw_data)
}
\arguments{
\item{start_flat_raw_data}{}
}
\value{

}
\description{
Flats raw data by selecting good known flag r and define Target variable.
}
