% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{data_tools.load_raw_data}
\alias{data_tools.load_raw_data}
\title{Loads all data needed for forecasting.}
\usage{
data_tools.load_raw_data(
  indice_code,
  day_date,
  geometric_point_t1 = F,
  geometric_point_t2 = F,
  delta_point = T,
  dynamic_flag = T
)
}
\arguments{
\item{indice_code}{a character vector indicating the indice code on which forecast will be done.}

\item{day_date}{a character vector indicating the date of the day on which forecast will be done.}
}
\value{
a list of \code{data.table} containing all useful data for the forecasting process.
}
\description{
Loads all data needed for forecasting.
}
