% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{data_tools.target_to_factor}
\alias{data_tools.target_to_factor}
\title{Title}
\usage{
data_tools.target_to_factor(learning_data, order, level)
}
\arguments{
\item{level}{}
}
\value{

}
\description{
Title
}
