% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{data_tools.update_raw_data}
\alias{data_tools.update_raw_data}
\title{Update previously computed data used for forecasting with new day_date data.}
\usage{
data_tools.update_raw_data(
  raw_data,
  indice_code,
  day_date,
  geometric_point_t1 = F,
  geometric_point_t2 = F,
  delta_point = T,
  dynamic_flag = T
)
}
\arguments{
\item{raw_data}{raw data initially coming from \code{data_tools.load_raw_data}}

\item{indice_code}{a character vector indicating the indice code on which forecast will be done.}

\item{day_date}{a character vector indicating the date of the day on which forecast will be done.}

\item{geometric_point_t1}{a boolean indicating if geometric point of type 1 should be computed}

\item{geometric_point_t2}{a boolean indicating if geometric point of type 2 should be computed}

\item{delta_point}{a boolean indicating if delta point should be computed}

\item{dynamic_flag}{a boolean indicating if dynamic flag should be computed}
}
\value{
a list of \code{data.table} containing all useful data for the forecasting process.
}
\description{
Update previously computed data used for forecasting with new day_date data.
}
